Domain field is not set in edit mode of multivisualisation wizard. Fi…

+++ abixen-platform-core/src/main/java/com/abixen/platform/core/service/impl/ModuleServiceImpl.java
@@ -120,8 +120,10 @@ public class ModuleServiceImpl implements ModuleService {
         List<Module> modules = moduleRepository.findAllExcept(page, ids);
 
         List<Long> moduleIds = modules.stream().map(module -> module.getId()).collect(Collectors.toList());
-        commentService.deleteCommentByModuleIds(moduleIds);
 
+        if (!moduleIds.isEmpty()) {
+            commentService.deleteCommentByModuleIds(moduleIds);
+        }
         moduleRepository.removeAllExcept(page, ids);
 
         modules.forEach(module -> {
